+++ pons/_client.py
@@ -352,6 +352,29 @@ class ClientSession:
         )
         return rpc_decode_quantity(result)
 
+    @rpc_call("eth_getCode")
+    async def eth_get_code(
+        self, address: Address, block: Union[int, Block] = Block.LATEST
+    ) -> bytes:
+        """Calls the ``eth_getCode`` RPC method."""
+        result = await self._provider_session.rpc(
+            "eth_getCode", address.rpc_encode(), rpc_encode_block(block)
+        )
+        return rpc_decode_data(result)
+
+    @rpc_call("eth_getStorageAt")
+    async def eth_get_storage_at(
+        self, address: Address, position: int, block: Union[int, Block] = Block.LATEST
+    ) -> bytes:
+        """Calls the ``eth_getCode`` RPC method."""
+        result = await self._provider_session.rpc(
+            "eth_getStorageAt",
+            address.rpc_encode(),
+            rpc_encode_quantity(position),
+            rpc_encode_block(block),
+        )
+        return rpc_decode_data(result)
+
     async def wait_for_transaction_receipt(
         self, tx_hash: TxHash, poll_latency: float = 1.0
     ) -> TxReceipt:
@@ -523,7 +546,7 @@ class ClientSession:
             gas = await self.estimate_transfer(signer.address, destination_address, amount)
         # TODO (#19): implement gas strategies
         max_gas_price = await self.eth_gas_price()
-        max_tip = Amount.gwei(1)
+        max_tip = min(Amount.gwei(1), max_gas_price)
         nonce = await self.eth_get_transaction_count(signer.address, Block.LATEST)
         tx: Dict[str, Union[int, str]] = {
             "type": 2,  # EIP-2930 transaction
@@ -589,7 +612,7 @@ class ClientSession:
             gas = await self.estimate_deploy(signer.address, call, amount=amount)
         # TODO (#19): implement gas strategies
         max_gas_price = await self.eth_gas_price()
-        max_tip = Amount.gwei(1)
+        max_tip = min(Amount.gwei(1), max_gas_price)
         nonce = await self.eth_get_transaction_count(signer.address, Block.LATEST)
         tx: Dict[str, Union[int, str]] = {
             "type": 2,  # EIP-2930 transaction
@@ -641,7 +664,7 @@ class ClientSession:
             gas = await self.estimate_transact(signer.address, call, amount=amount)
         # TODO (#19): implement gas strategies
         max_gas_price = await self.eth_gas_price()
-        max_tip = Amount.gwei(1)
+        max_tip = min(Amount.gwei(1), max_gas_price)
         nonce = await self.eth_get_transaction_count(signer.address, Block.LATEST)
         tx: Dict[str, Union[int, str]] = {
             "type": 2,  # EIP-2930 transaction